Automate instantiation of memory evars
Lithium only instantiates (via unification) existential variables
(evars) that are an equality like `evar = x`. Rewrite an evar equation
in the reading and writing memory Lithium wp rules such that the many
manual evar instantiations with the liInst tactic are not required
anymore.

This adds an additional proof obligation: Showing that the number of
bytes of an array element divides the address offset. This is always
provable when using the pre-defined definition for address points to
array. They should be provable using

  `apply Z.mod_divide; [lia|bv_solve].`